Refactoring #26: Fixing boxes, take five! (Or is it six?)
At this point I kind of gave up:
"Let's just make the middle line two characters wider than we think it should be (62) and see how that looks."

Thankfully, the AI was eager to please:
"Good idea! Let's make the middle lines 62 characters. Using the formula with 58 instead of 56..."

And now it looks much better.